tk ui: prevent empty module creation

+++ lib/App/CPAN2Pkg/UI/Tk.pm
@@ -285,6 +285,7 @@ event _on_btn_submit => sub {
     my $self = shift;
     my $entry = $self->_w( "ent_module" );
     my $module = $entry->get;
+    return unless $module;
     $K->post( controller => new_module_wanted => $module );
     $entry->delete( 0, 'end' );
 };
